slå sammen kode
