+++ src/main/kotlin/no/nav/personbruker/dittnav/eventhandler/done/Done.kt
@@ -1,5 +1,7 @@
 package no.nav.personbruker.dittnav.eventhandler.done
 
+
+import kotlinx.serialization.Serializable
 import java.time.ZonedDateTime
 
 data class Done(
@@ -24,3 +26,9 @@ data class Done(
             "forstBehandlet=$forstBehandlet"
     }
 }
+
+
+@Serializable
+data class DoneDTO(
+    val eventId: String
+)
