--- src/main/kotlin/no/nav/personbruker/dittnav/eventhandler/done/DoneDTO.kt
@@ -1,8 +1,0 @@
-package no.nav.personbruker.dittnav.eventhandler.done
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class DoneDTO(
-    val eventId: String
-)
